add scan_force func for 'scan' command. A bit of visual update

+++ src/commands.rs
@@ -6,7 +6,7 @@ use crate::{color, comments, config, parser};
 
 fn init_first() {
     let error = color::paint_str("Error:".to_string(), color::Color::Red);
-    let chant_init = color::paint_str("'$ chant init'".to_string(), color::Color::Yellow);
+    let chant_init = color::paint_str("$chant init".to_string(), color::Color::Yellow);
     println!("{error} chant wasn't initialised. Run {chant_init}");
 }
 
@@ -20,7 +20,7 @@ pub fn init(){
         config::create_config();
     } else {
         let error = color::paint_str("Error:".to_string(), color::Color::Red);
-        let chant_dismiss = color::paint_str("$ chant dismiss".to_string(), color::Color::Yellow);
+        let chant_dismiss = color::paint_str("$chant dismiss".to_string(), color::Color::Yellow);
         println!("{error} Chant is already initialised.\nRun {chant_dismiss} remove chant from this directory");
         return;
     }
@@ -29,7 +29,7 @@ pub fn init(){
 
     let chant = color::paint_str("Chant".to_string(), color::Color::Cyan);
 
-    scan();
+    scan_force();
 
     println!("{chant} was initialised successfully");
 }
@@ -50,6 +50,38 @@ pub fn bad_syntax() {
 pub fn unknown_command() {
     let error = color::paint_str("Error:".to_string(), color::Color::Red);
     println!("{error} unknown command");
+}
+
+pub fn scan_force() {
+    if !is_initialised(){
+        init_first();
+        return; 
+    }
+
+    let config = config::read_config();
+
+    let mut new_storage = comments::new_storage();
+
+    let mut it = WalkDir::new(".").into_iter();
+    while let Some(entry) = it.next(){
+        match entry {
+            Ok(e) => {
+                let file_name = e.file_name().to_string_lossy();
+                let path = &e.path().to_string_lossy();
+                if e.file_type().is_dir() && config.ignore.contains(&file_name.to_string()) {
+                    it.skip_current_dir();
+                    continue;
+                } else if !e.file_type().is_dir() && let Some(ext) = e.path().extension() {
+                    if config.read.contains(&ext.display().to_string()){
+                        let new_file = parser::parse_new_file(path.to_string());
+                        new_storage.files.insert(path.to_string(), new_file);
+                    }
+                }
+            },
+            Err(e) => println!("unable to scan: {}", e),
+        }
+    }
+    save_storage(&new_storage);
 }
 
 pub fn scan() {
@@ -105,6 +137,9 @@ pub fn list() {
     }
 
     for file in s.files {
+        if file.1.comments.len() > 0 {
+            println!();
+        }
         for comment in file.1.comments {
             let c = comment.1;
 
@@ -121,6 +156,7 @@ pub fn list() {
             println!("[{}] {}:{} - {}", kind, path, index, c.line);
         }
     }
+    println!();
 }
 
 pub fn dismiss() {
@@ -134,34 +170,46 @@ pub fn dismiss() {
     println!("{chant} was removed");
 }
 
+fn is_gitignore_exists() -> bool {
+    let ex = fs::exists("./.gitignore");
+    match ex {
+        Ok(v)=> return v,
+        Err(e)=> {println!("{}", e); return false;}
+    }
+}
+
 fn add_to_gitignore() {
-    let content = fs::read_to_string(".gitignore");
-    match content {
-        Ok(v) => {
-            let ignore = format!("{}\n.chant", v);
-        let _ = fs::write(".gitignore", ignore);
-    },
-        Err(_) => { return }
+    if is_gitignore_exists() {
+        let content = fs::read_to_string(".gitignore");
+        match content {
+            Ok(v) => {
+                let ignore = format!("{}\n.chant", v);
+            let _ = fs::write(".gitignore", ignore);
+        },
+            Err(_) => { return }
+        }
     }
 }
 
 fn remove_from_gitignore() {
-    let content = fs::read_to_string(".gitignore");
-    match content {
-        Ok(v) => {
-            let lines = v.split("\n");
-            let mut ignore = "".to_owned();
-            for line in lines {
-                if line != ".chant" {
-                    ignore.push_str(line); 
+    if is_gitignore_exists() {
+        let content = fs::read_to_string(".gitignore");
+        match content {
+            Ok(v) => {
+                let lines = v.split("\n");
+                let mut ignore = "".to_owned();
+                for line in lines {
+                    if line != ".chant" {
+                        ignore.push_str(line); 
+                    }
                 }
+                let _ = fs::write(".gitignore", ignore);
+            },
+            Err(e) => {
+                let error = color::paint_str("Error:".to_string(), color::Color::Red);
+                println!("{error} unable to remove .chant/ direcotry from .gitignore. You should do this manually\n{e}");
+                return
             }
-            let _ = fs::write(".gitignore", ignore);
-        },
-        Err(e) => {
-            let error = color::paint_str("Error:".to_string(), color::Color::Red);
-            println!("{error} unable to remove .chant/ direcotry from .gitignore. You should do this manually\n{e}");
-            return
         }
     }
 }
